more mocks and stubs moved to testscommon
