--- integrationTests/mock/txTypeHandlerMock.go
@@ -1,25 +1,0 @@
-package mock
-
-import (
-	"github.com/ElrondNetwork/elrond-go/data"
-	"github.com/ElrondNetwork/elrond-go/process"
-)
-
-// TxTypeHandlerMock -
-type TxTypeHandlerMock struct {
-	ComputeTransactionTypeCalled func(tx data.TransactionHandler) (process.TransactionType, process.TransactionType)
-}
-
-// ComputeTransactionType -
-func (th *TxTypeHandlerMock) ComputeTransactionType(tx data.TransactionHandler) (process.TransactionType, process.TransactionType) {
-	if th.ComputeTransactionTypeCalled == nil {
-		return process.MoveBalance, process.MoveBalance
-	}
-
-	return th.ComputeTransactionTypeCalled(tx)
-}
-
-// IsInterfaceNil -
-func (th *TxTypeHandlerMock) IsInterfaceNil() bool {
-	return th == nil
-}
