+++ integrationTests/vm/arwen/arwenvm/arwenVM_test.go
@@ -524,7 +524,7 @@ func TestExecuteTransactionAndTimeToProcessChange(t *testing.T) {
 		SignMarshalizer:  testMarshalizer,
 		ShardCoordinator: shardCoordinator,
 		ScProcessor:      &testscommon.SCProcessorMock{},
-		TxFeeHandler:     &mock.UnsignedTxHandlerMock{},
+		TxFeeHandler:     &testscommon.UnsignedTxHandlerMock{},
 		TxTypeHandler:    txTypeHandler,
 		EconomicsFee:     &mock.FeeHandlerStub{},
 		ReceiptForwarder: &mock.IntermediateTransactionHandlerMock{},
