+++ integrationTests/vm/arwen/utils.go
@@ -157,7 +157,7 @@ func SetupTestContext(t *testing.T) *TestContext {
 }
 
 func (context *TestContext) initFeeHandlers() {
-	context.UnsignexTxHandler = &mock.UnsignedTxHandlerMock{
+	context.UnsignexTxHandler = &testscommon.UnsignedTxHandlerMock{
 		ProcessTransactionFeeCalled: func(cost *big.Int, hash []byte) {
 			context.LastConsumedFee = cost.Uint64()
 		},
