+++ integrationTests/vm/testInitializer.go
@@ -448,7 +448,7 @@ func CreateTxProcessorWithOneSCExecutorMockVM(
 		ShardCoordinator: oneShardCoordinator,
 		ScrForwarder:     &mock.IntermediateTransactionHandlerMock{},
 		BadTxForwarder:   &mock.IntermediateTransactionHandlerMock{},
-		TxFeeHandler:     &mock.UnsignedTxHandlerMock{},
+		TxFeeHandler:     &testscommon.UnsignedTxHandlerMock{},
 		EconomicsFee:     economicsData,
 		TxTypeHandler:    txTypeHandler,
 		GasHandler: &mock.GasHandlerMock{
@@ -474,7 +474,7 @@ func CreateTxProcessorWithOneSCExecutorMockVM(
 		SignMarshalizer:                testMarshalizer,
 		ShardCoordinator:               oneShardCoordinator,
 		ScProcessor:                    scProcessor,
-		TxFeeHandler:                   &mock.UnsignedTxHandlerMock{},
+		TxFeeHandler:                   &testscommon.UnsignedTxHandlerMock{},
 		TxTypeHandler:                  txTypeHandler,
 		EconomicsFee:                   economicsData,
 		ReceiptForwarder:               &mock.IntermediateTransactionHandlerMock{},
