+++ testscommon/txTypeHandlerMock.go
@@ -1,4 +1,4 @@
-package mock
+package testscommon
 
 import (
 	"github.com/ElrondNetwork/elrond-go/data"
@@ -19,7 +19,7 @@ func (th *TxTypeHandlerMock) ComputeTransactionType(tx data.TransactionHandler) 
 	return th.ComputeTransactionTypeCalled(tx)
 }
 
-// IsInterfaceNil returns true if there is no value under the interface
+// IsInterfaceNil -
 func (th *TxTypeHandlerMock) IsInterfaceNil() bool {
 	return th == nil
 }
